Plex RAW photos/picture support
This feature was requested. I got it implemented by passing the path of
the image, if it is not a normal image file, through the
getTranscodeImagePath function in the PlexAPI class. This requests Plex
to transcode the file before sending it to Kodi. Slight disadvantage is
the maximum resolution of 1920x1080, but I doubt that will impact many
people.